fix: login and register success flash messages Due to page refresh, they cannot show the alert component (the dialog instantly closes).
Toast wouldn't show neither, since the flash messages are only available to the next request, so calling `toast` with the current flash will fail (flash messages doesn't exist yet).

Then, once the props change, a useEffect watch to it and check if either one of the flash messages are available. If so, it displays them and then delete them so that they are not shown twice.